requirements.txt
Flask
pandas
scikit-learn==1.3.0
joblib
dnspython
requests
python-whois
